%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{[.RspDocument}
\alias{[.RspDocument}
\alias{RspDocument.[}
\alias{[,RspDocument-method}
\alias{RspDocument.[<-}
\alias{[<-.RspDocument}
\alias{[<-,RspDocument-method}


\title{Subsets an RspDocument}

\description{
 Subsets an RspDocument.
}

\usage{
\method{[}{RspDocument}(x, i)
}

\arguments{
  \item{i}{Indices of the RSP elements to extract.}
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspDocument}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
