%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.HttpDaemon}
\alias{as.character.HttpDaemon}
\alias{HttpDaemon.as.character}
\alias{as.character,HttpDaemon-method}

\title{Returns a short string describing the HTTP daemon}

\description{
 Returns a short string describing the HTTP daemon.
}

\usage{
\method{as.character}{HttpDaemon}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
