%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.HttpRequest}
\alias{as.character.HttpRequest}
\alias{HttpRequest.as.character}
\alias{as.character.HttpRequest}
\alias{as.character,HttpRequest-method}


\title{Returns a short string describing the HTTP request}

\description{
 Returns a short string describing the HTTP request.
}

\usage{\method{as.character}{HttpRequest}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpRequest}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
