%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspStringProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.RspStringProduct}
\alias{as.character.RspStringProduct}
\alias{RspStringProduct.as.character}
\alias{as.character,RspStringProduct-method}

\title{Returns a plain character string representation of an RSP string product}

\description{
 Returns a plain character string representation of an RSP string product.  All attributes including class have been dropped.
}

\usage{
\method{as.character}{RspStringProduct}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspStringProduct}}.
}
\keyword{internal}
\keyword{methods}
