%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspCode}
\alias{asRspString.RspCode}
\alias{RspCode.asRspString}
\alias{asRspString,RspCode-method}

\title{Recreates an RSP string from an RspCode}

\description{
 Recreates an RSP string from an RspCode.
}

\usage{
\method{asRspString}{RspCode}(code, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspCode}}.
}
\keyword{internal}
\keyword{methods}
