%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspCodeChunk}
\alias{asRspString.RspCodeChunk}
\alias{RspCodeChunk.asRspString}
\alias{asRspString.RspCodeChunk}
\alias{asRspString,RspCodeChunk-method}


\title{Recreates an RSP string from an RspCodeChunk}

\description{
 Recreates an RSP string from an RspCodeChunk.
}

\usage{
\method{asRspString}{RspCodeChunk}(code, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspCodeChunk}}.
}
\keyword{internal}
\keyword{methods}
