%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspConstruct}
\alias{asRspString.RspConstruct}
\alias{RspConstruct.asRspString}
\alias{asRspString.RspConstruct}
\alias{asRspString,RspConstruct-method}


\title{Recreates an RSP string from an RspConstruct}

\description{
 Recreates an RSP string from an RspConstruct.
}

\usage{
\method{asRspString}{RspConstruct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
