%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspDirective}
\alias{asRspString.RspDirective}
\alias{RspDirective.asRspString}
\alias{asRspString,RspDirective-method}

\title{Recreates an RSP string from an RSP preprocessing directive}

\description{
 Recreates an RSP string from an RSP preprocessing directive.
}

\usage{
\method{asRspString}{RspDirective}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDirective}}.
}
\keyword{internal}
\keyword{methods}
