%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspDocument}
\alias{asRspString.RspDocument}
\alias{RspDocument.asRspString}
\alias{asRspString.RspDocument}
\alias{asRspString,RspDocument-method}


\title{Recreates an RSP string from an RspDocument}

\description{
 Recreates an RSP string from an RspDocument.
}

\usage{\method{asRspString}{RspDocument}(doc, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
