%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspText.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspText}
\alias{asRspString.RspText}
\alias{RspText.asRspString}
\alias{asRspString,RspText-method}

\title{Recreates an RSP string from an RspText}

\description{
 Recreates an RSP string from an RspText.
}

\usage{
\method{asRspString}{RspText}(text, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspText}}.
}
\keyword{internal}
\keyword{methods}
