%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspUnparsedDirective}
\alias{asRspString.RspUnparsedDirective}
\alias{RspUnparsedDirective.asRspString}
\alias{asRspString,RspUnparsedDirective-method}

\title{Recreates an RSP string from an unparsed RSP preprocessing directive}

\description{
 Recreates an RSP string from an unparsed RSP preprocessing directive.
}

\usage{
\method{asRspString}{RspUnparsedDirective}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspUnparsedDirective}}.
}
\keyword{internal}
\keyword{methods}
