%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  buildNonSweaveVignette.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{buildNonSweaveVignettes}
\alias{buildNonSweaveVignettes}


\title{Builds all non-Sweave Rnw vignette}

\description{
 Builds all non-Sweave Rnw vignette.
}

\usage{
buildNonSweaveVignettes(...)
}

\arguments{
  \item{...}{Additional arguments passed to \code{\link{buildNonSweaveVignette}}().}
}

\value{
  Returns (invisibly) a named \code{\link[base]{list}} with elements of what
  the vignette builder returns.
}

\author{Henrik Bengtsson}

\seealso{
  To build vignette, see \code{\link{buildNonSweaveVignette}}().
}




\keyword{file}
\keyword{IO}
\keyword{internal}
