%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  buildNonSweaveVignette.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{buildPkgIndexHtml}
\alias{buildPkgIndexHtml}


\title{Builds a package index HTML file}

\description{
 Builds a package index HTML file, iff missing.
}

\usage{
buildPkgIndexHtml(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) the absolute pathname to the built index.html file.
  If an index.html file already exists, nothing is done and \code{\link[base]{NULL}}
  is returned.
}

\author{Henrik Bengtsson}




\keyword{file}
\keyword{IO}
\keyword{internal}
