%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  buildVignette.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{buildVignette}
\alias{buildVignette}


\title{Builds a vignette}

\description{
 Builds a vignette into a PDF or a HTML file.
}

\usage{
buildVignette(file, dir=".", latex=TRUE, tangle=TRUE, quiet=TRUE, clean=TRUE, ...)
}

\arguments{
  \item{file}{A character string specifying the vignette source file.}
  \item{dir}{A character string specifying the working directory.}
  \item{latex}{A logical specifying whether a TeX produced by weave
    should be compiled by \link{texi2pdf} or not.}
  \item{tangle}{A logical specifying whether tangle should be run or not.}
  \item{quiet}{A logical specifying whether weave, \link{texi2pdf}, and
    tangle should be run in quiet mode or not.}
  \item{clean}{Remove all files generated by the build, even if there
    were copies there before.}
  \item{...}{Additional arguments passed to weave and tangle.}
}

\value{
  Returns the filename to the final vignette (PDF or HTML) product.
  If \code{latex = FALSE}, it may be a TeX file.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}




\keyword{file}
\keyword{IO}
\keyword{internal}
