%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{dropComments.RspString}
\alias{dropComments.RspString}
\alias{RspString.dropComments}
\alias{dropComments.RspString}
\alias{dropComments,RspString-method}


\title{Drops all RSP comments}

\description{
 Drops all RSP comments.
}

\usage{\method{dropComments}{RspString}(object, envir=parent.frame(), ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}} without the RSP comments.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspString}}.
}
\keyword{internal}
\keyword{methods}
