%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlRspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{escape.HtmlRspLanguage}
\alias{escape.HtmlRspLanguage}
\alias{HtmlRspLanguage.escape}
\alias{escape.HtmlRspLanguage}
\alias{escape,HtmlRspLanguage-method}


\title{Escapes a string specifically for the HTML language}

\description{
 Escapes a string specifically for the HTML language.
}

\usage{\method{escape}{HtmlRspLanguage}(object, ...)}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HtmlRspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
