%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{escape.RspLanguage}
\alias{escape.RspLanguage}
\alias{RspLanguage.escape}
\alias{escape,RspLanguage-method}

\title{Escapes a string specifically for a given RSP response language}

\description{
 Escapes a string specifically for a given RSP response language.
}

\usage{
\method{escape}{RspLanguage}(this, ...)
}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
