%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.toR.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{evaluate.RspDocument}
\alias{evaluate.RspDocument}
\alias{RspDocument.evaluate}
\alias{evaluate,RspDocument-method}

\title{Parses, translates, and evaluates the RSP document}

\description{
 Parses, translates, and evaluates the RSP document.
}

\usage{
\method{evaluate}{RspDocument}(object, envir=parent.frame(), ...)
}

\arguments{
  \item{envir}{The \code{\link[base]{environment}} where the RSP document is evaluated.}
  \item{...}{Not used.}
}

\value{
 Returns the last evaluated expression, iff any.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
