%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{evaluate.RspSourceCode}
\alias{evaluate.RspSourceCode}
\alias{RspSourceCode.evaluate}
\alias{evaluate.RspSourceCode}
\alias{evaluate,RspSourceCode-method}


\title{Parses and evaluates the source code}

\description{
 Parses and evaluates the source code.
}

\usage{\method{evaluate}{RspSourceCode}(...)}

\arguments{
  \item{...}{Optional arguments passed to \code{\link[R.rsp:parse.RspSourceCode]{*parse}()}.}
}

\value{
 Returns the last evaluated expression, iff any.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspSourceCode}}.
}
\keyword{internal}
\keyword{methods}
