%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspString.toR.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{evaluate.RspString}
\alias{evaluate.RspString}
\alias{RspString.evaluate}
\alias{evaluate,RspString-method}

\title{Parses, translates, and evaluates the RSP string}

\description{
 Parses, translates, and evaluates the RSP string.
}

\usage{
\method{evaluate}{RspString}(object, envir=parent.frame(), ...)
}

\arguments{
  \item{envir}{The \code{\link[base]{environment}} where the RSP string is evaluated.}
  \item{...}{Not used.}
}

\value{
 Returns the last evaluated expression, iff any.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspString}}.
}
\keyword{internal}
\keyword{methods}
