%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{evaluate.SourceCode}
\alias{evaluate.SourceCode}
\alias{SourceCode.evaluate}
\alias{evaluate.SourceCode}
\alias{evaluate,SourceCode-method}


\title{Parses and evaluates the source code}

\description{
 Parses and evaluates the source code.
}

\usage{\method{evaluate}{SourceCode}(...)}

\arguments{
  \item{...}{Optional arguments passed to \code{\link[R.rsp:parse.SourceCode]{*parse}()}.}
}

\value{
 Returns the last evaluated expression, iff any.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SourceCode}}.
}
\keyword{internal}
\keyword{methods}
