%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspBashSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{exprToCode.RspBashSourceCodeFactory}
\alias{exprToCode.RspBashSourceCodeFactory}
\alias{RspBashSourceCodeFactory.exprToCode}
\alias{exprToCode,RspBashSourceCodeFactory-method}

\title{Translates an RspExpression into bash source code}

\description{
 Translates an RspExpression into bash source code.
}

\usage{
\method{exprToCode}{RspBashSourceCodeFactory}(object, expr, ..., index=NA)
}

\arguments{
  \item{expr}{An \code{\link{RspExpression}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspBashSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
