%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{exprToCode.RspRSourceCodeFactory}
\alias{exprToCode.RspRSourceCodeFactory}
\alias{RspRSourceCodeFactory.exprToCode}
\alias{exprToCode,RspRSourceCodeFactory-method}

\title{Translates an RspExpression into R source code}

\description{
 Translates an RspExpression into R source code.
}

\usage{
\method{exprToCode}{RspRSourceCodeFactory}(object, expr, ..., index=NA)
}

\arguments{
  \item{expr}{An \code{\link{RspExpression}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspRSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
