%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{exprToCode.RspSourceCodeFactory}
\alias{exprToCode.RspSourceCodeFactory}
\alias{RspSourceCodeFactory.exprToCode}
\alias{exprToCode.RspSourceCodeFactory}
\alias{exprToCode,RspSourceCodeFactory-method}


\title{Translates an RspExpression into source code}

\description{
 Translates an RspExpression into source code.
}

\usage{\method{exprToCode}{RspSourceCodeFactory}(...)}

\arguments{
  \item{expr}{An \code{\link{RspExpression}}.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
