%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{flush.FileRspResponse}
\alias{flush.FileRspResponse}
\alias{FileRspResponse.flush}
\alias{flush.FileRspResponse}
\alias{flush,FileRspResponse-method}


\title{Flushes the response buffer}

\description{
 Flushes the response buffer.
}

\usage{\method{flush}{FileRspResponse}(con, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FileRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
