%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemonRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{flush.HttpDaemonRspResponse}
\alias{flush.HttpDaemonRspResponse}
\alias{HttpDaemonRspResponse.flush}
\alias{flush.HttpDaemonRspResponse}
\alias{flush,HttpDaemonRspResponse-method}


\title{Flushes the buffer of an HttpDaemonRspResponse to the HttpDaemon}
\description{
 Flushes the buffer of an HttpDaemonRspResponse to the HttpDaemon.
}

\usage{\method{flush}{HttpDaemonRspResponse}(con, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns (invisibly) the number of characters flushed.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemonRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
