%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{flush.RspResponse}
\alias{flush.RspResponse}
\alias{RspResponse.flush}
\alias{flush.RspResponse}
\alias{flush,RspResponse-method}


\title{Flushes the response buffer}

\description{
 Flushes the response buffer.
}

\usage{\method{flush}{RspResponse}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
