%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAbsolutePath.FileRspResponse}
\alias{getAbsolutePath.FileRspResponse}
\alias{FileRspResponse.getAbsolutePath}
\alias{getAbsolutePath.FileRspResponse}
\alias{getAbsolutePath,FileRspResponse-method}


\title{Gets the absolute pathname to the current RSP file}

\description{
 Gets the absolute pathname to the current RSP file.
}

\usage{\method{getAbsolutePath}{FileRspResponse}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FileRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
