%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspPage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAbsolutePath.RspPage}
\alias{getAbsolutePath.RspPage}
\alias{RspPage.getAbsolutePath}
\alias{getAbsolutePath.RspPage}
\alias{getAbsolutePath,RspPage-method}


\title{Gets the absolute pathname to the current RSP file}

\description{
 Gets the absolute pathname to the current RSP file.
}

\usage{\method{getAbsolutePath}{RspPage}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspPage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
