%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAnnotation.RspDocument}
\alias{getAnnotation.RspDocument}
\alias{RspDocument.getAnnotation}
\alias{getAnnotation.RspDocument}
\alias{getAnnotation,RspDocument-method}


\title{Gets the annotation of the RspDocument}

\description{
 Gets the annotation of the RspDocument.
}

\usage{
\method{getAnnotation}{RspDocument}(object, name=NULL, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
