%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAnnotation.RspProduct}
\alias{getAnnotation.RspProduct}
\alias{RspProduct.getAnnotation}
\alias{getAnnotation.RspProduct}
\alias{getAnnotation,RspProduct-method}


\title{Gets the annotation an RSP product}

\description{
 Gets the annotation an RSP product.
}

\usage{
\method{getAnnotation}{RspProduct}(object, name=NULL, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspProduct}}.
}
\keyword{internal}
\keyword{methods}
