%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAttributes.RspDocument}
\alias{getAttributes.RspDocument}
\alias{RspDocument.getAttributes}
\alias{getAttributes,RspDocument-method}
\alias{RspDocument.getAttribute}
\alias{getAttribute.RspDocument}
\alias{getAttribute,RspDocument-method}

\alias{RspDocument.setAttributes}
\alias{setAttributes.RspDocument}
\alias{setAttributes,RspDocument-method}

\alias{RspDocument.setAttribute}
\alias{setAttribute.RspDocument}
\alias{setAttribute,RspDocument-method}


\title{Gets the attributes of an RSP document}

\description{
 Gets the attributes of an RSP document.
}

\usage{
\method{getAttributes}{RspDocument}(object, private=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
