%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspExpression.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAttributes.RspExpression}
\alias{getAttributes.RspExpression}
\alias{RspExpression.getAttributes}
\alias{getAttributes.RspExpression}
\alias{getAttributes,RspExpression-method}


\title{Gets the attributes of an RSP expression}

\description{
 Gets the attributes of an RSP expression.
}

\usage{\method{getAttributes}{RspExpression}(directive, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspExpression}}.
}
\keyword{internal}
\keyword{methods}
