%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAttributes.RspObject}
\alias{getAttributes.RspObject}
\alias{RspObject.getAttributes}
\alias{getAttributes,RspObject-method}
\alias{RspObject.getAttribute}
\alias{getAttribute.RspObject}
\alias{getAttribute,RspObject-method}

\alias{RspObject.hasAttribute}
\alias{hasAttribute.RspObject}
\alias{hasAttribute,RspObject-method}

\alias{RspObject.setAttributes}
\alias{setAttributes.RspObject}
\alias{setAttributes,RspObject-method}

\alias{RspObject.setAttribute}
\alias{setAttribute.RspObject}
\alias{setAttribute,RspObject-method}


\title{Gets the attributes of an RSP object}

\description{
 Gets the attributes of an RSP object.
}

\usage{
\method{getAttributes}{RspObject}(object, private=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspObject}}.
}
\keyword{internal}
\keyword{methods}
