%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAttributes.RspString}
\alias{getAttributes.RspString}
\alias{RspString.getAttributes}
\alias{getAttributes,RspString-method}
\alias{RspString.getAttribute}
\alias{getAttribute.RspString}
\alias{getAttribute,RspString-method}

\alias{RspString.setAttributes}
\alias{setAttributes.RspString}
\alias{setAttributes,RspString-method}

\alias{RspString.setAttribute}
\alias{setAttribute.RspString}
\alias{setAttribute,RspString-method}


\title{Gets the attributes of an RSP document}

\description{
 Gets the attributes of an RSP document.
}

\usage{
\method{getAttributes}{RspString}(object, private=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspString}}.
}
\keyword{internal}
\keyword{methods}
