%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlRspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getComment.HtmlRspLanguage}
\alias{getComment.HtmlRspLanguage}
\alias{HtmlRspLanguage.getComment}
\alias{getComment.HtmlRspLanguage}
\alias{getComment,HtmlRspLanguage-method}


\title{Gets a comment string specifically for the HTML language}

\description{
 Gets a comment string specifically for the HTML language.
}

\usage{\method{getComment}{HtmlRspLanguage}(object, ...)}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HtmlRspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
