%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getComment.RspConstruct}
\alias{getComment.RspConstruct}
\alias{RspConstruct.getComment}
\alias{getComment,RspConstruct-method}

\title{Gets the comment of an RSP construct}

\description{
 Gets the comment of an RSP construct.
}

\usage{
\method{getComment}{RspConstruct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
