%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getComment.RspLanguage}
\alias{getComment.RspLanguage}
\alias{RspLanguage.getComment}
\alias{getComment,RspLanguage-method}
\alias{getComment.HtmlRspLanguage}

\title{Gets a comment string specific for a given RSP response language}

\description{
 Gets a comment string specific for a given RSP response language.
}

\usage{
\method{getComment}{RspLanguage}(this, ...)
}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
