%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCompleteCode.RspRSourceCode}
\alias{getCompleteCode.RspRSourceCode}
\alias{RspRSourceCode.getCompleteCode}
\alias{getCompleteCode,RspRSourceCode-method}

\title{Gets the complete R source code}

\description{
 Gets the complete R source code with output functions defined.
}

\usage{
\method{getCompleteCode}{RspRSourceCode}(object, output=c("stdout", "string"), ...)
}

\arguments{
  \item{output}{A \code{\link[base]{character}} string specifying type of output function.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspRSourceCode}}.
}
\keyword{internal}
\keyword{methods}
