%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCompleteCode.RspRSourceCodeFactory}
\alias{getCompleteCode.RspRSourceCodeFactory}
\alias{RspRSourceCodeFactory.getCompleteCode}
\alias{getCompleteCode,RspRSourceCodeFactory-method}

\title{Gets the complete R source code}

\description{
 Gets the complete R source code with output functions defined.
}

\usage{
\method{getCompleteCode}{RspRSourceCodeFactory}(this, object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspRSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
