%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCompleteCode.RspSourceCodeFactory}
\alias{getCompleteCode.RspSourceCodeFactory}
\alias{RspSourceCodeFactory.getCompleteCode}
\alias{getCompleteCode,RspSourceCodeFactory-method}
\alias{getCompleteCode.RspRSourceCodeFactory}

\title{Gets the source code header, body, and footer}

\description{
 Gets the source code header, body, and footer.
}

\usage{
\method{getCompleteCode}{RspSourceCodeFactory}(this, object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
