%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$getConfig}
\alias{HttpDaemon$getConfig}
\alias{getConfig.HttpDaemon}
\alias{HttpDaemon.getConfig}
\alias{getConfig,HttpDaemon-method}

\title{Retrieves the server's 'config' structure from Tcl}

\description{
 Retrieves the server's 'config' structure from Tcl.
}

\usage{
## Static method (use this):
## HttpDaemon$getConfig(...)

## Don't use the below:
\method{getConfig}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a tclArray object.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
