%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getContent.RspEvalDirective}
\alias{getContent.RspEvalDirective}
\alias{RspEvalDirective.getContent}
\alias{getContent,RspEvalDirective-method}

\title{Gets the content of the RSP eval directive}

\description{
 Gets the content of the RSP eval directive.
}

\usage{
\method{getContent}{RspEvalDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspEvalDirective}}.
}
\keyword{internal}
\keyword{methods}
