%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getContent.RspIncludeDirective}
\alias{getContent.RspIncludeDirective}
\alias{RspIncludeDirective.getContent}
\alias{getContent,RspIncludeDirective-method}

\title{Gets the content of the RSP include directive}

\description{
 Gets the content of the RSP include directive.
}

\usage{
\method{getContent}{RspIncludeDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspIncludeDirective}}.
}
\keyword{internal}
\keyword{methods}
