%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getContentType.HttpRequest}
\alias{getContentType.HttpRequest}
\alias{HttpRequest.getContentType}
\alias{getContentType,HttpRequest-method}

\title{Gets the MIME type of the body of the request}

\description{
 Gets the MIME type of the body of the request, or \code{\link[base]{NULL}} if the type is not known.
}

\usage{
\method{getContentType}{HttpRequest}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
