%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDefaultFilename.HttpDaemon}
\alias{getDefaultFilename.HttpDaemon}
\alias{HttpDaemon.getDefaultFilename}
\alias{getDefaultFilename.HttpDaemon}
\alias{getDefaultFilename,HttpDaemon-method}


\title{Gets the default filename to be loaded by the HTTP daemon}

\description{
 Gets the default filename to be loaded by the HTTP daemon, if started.
}

\synopsis{getDefaultFilename.HttpDaemon(static, ...)}
\usage{HttpDaemon$getDefaultFilename(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{character}} string if started, otherwise \code{\link[base]{NA}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
