%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$getDefaultFilenamePattern}
\alias{HttpDaemon$getDefaultFilenamePattern}
\alias{getDefaultFilenamePattern.HttpDaemon}
\alias{HttpDaemon.getDefaultFilenamePattern}
\alias{getDefaultFilenamePattern,HttpDaemon-method}

\title{Gets the default filename pattern to be loaded by the HTTP daemon}

\description{
 Gets the default filename pattern to be loaded by the HTTP daemon, if started.
}

\usage{
## Static method (use this):
## HttpDaemon$getDefaultFilenamePattern(...)

## Don't use the below:
\method{getDefaultFilenamePattern}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{character}} string if started, otherwise \code{\link[base]{NA}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
