%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getEcho.RspCode}
\alias{getEcho.RspCode}
\alias{RspCode.getEcho}
\alias{getEcho,RspCode-method}

\title{Checks whether the source code should be echoed or not}

\description{
 Checks whether the source code should be echoed or not.
}

\usage{
\method{getEcho}{RspCode}(code, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspCode}}.
}
\keyword{internal}
\keyword{methods}
