%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFile.RspEvalDirective}
\alias{getFile.RspEvalDirective}
\alias{RspEvalDirective.getFile}
\alias{getFile,RspEvalDirective-method}

\title{Gets the file attribute}

\description{
 Gets the file attribute.
}

\usage{
\method{getFile}{RspEvalDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspEvalDirective}}.
}
\keyword{internal}
\keyword{methods}
