%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspExpression.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFile.RspIncludeDirective}
\alias{getFile.RspIncludeDirective}
\alias{RspIncludeDirective.getFile}
\alias{getFile.RspIncludeDirective}
\alias{getFile,RspIncludeDirective-method}


\title{Gets the file attribute}

\description{
 Gets the file attribute.
}

\usage{\method{getFile}{RspIncludeDirective}(directive, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspIncludeDirective}}.
}
\keyword{internal}
\keyword{methods}
