%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$getHttpRequest}
\alias{HttpDaemon$getHttpRequest}
\alias{getHttpRequest.HttpDaemon}
\alias{HttpDaemon.getHttpRequest}
\alias{getHttpRequest,HttpDaemon-method}

\title{Gets the HTTP request}

\description{
 Gets the HTTP request.
}

\usage{
## Static method (use this):
## HttpDaemon$getHttpRequest(...)

## Don't use the below:
\method{getHttpRequest}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{HttpRequest}} object.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
