%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInclude.RspConstruct}
\alias{getInclude.RspConstruct}
\alias{RspConstruct.getInclude}
\alias{getInclude,RspConstruct-method}
\alias{getInclude.RspText}
\alias{getInclude.RspCodeChunk}
\alias{getInclude.RspVariableDirective}

\title{Checks whether an RSP construct will include text to the output or not}

\description{
 Checks whether an RSP construct will include text to the output or not.
}

\usage{
\method{getInclude}{RspConstruct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} of \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
