%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getKeywords.RspPageDirective}
\alias{getKeywords.RspPageDirective}
\alias{RspPageDirective.getKeywords}
\alias{getKeywords.RspPageDirective}
\alias{getKeywords,RspPageDirective-method}


\title{Gets the keywords}

\description{
 Gets the keywords.
}

\usage{
\method{getKeywords}{RspPageDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspPageDirective}}.
}
\keyword{internal}
\keyword{methods}
