%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLanguage.RspEvalDirective}
\alias{getLanguage.RspEvalDirective}
\alias{RspEvalDirective.getLanguage}
\alias{getLanguage,RspEvalDirective-method}

\title{Gets the programming language}

\description{
 Gets the programming language.
}

\usage{
\method{getLanguage}{RspEvalDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspEvalDirective}}.
}
\keyword{internal}
\keyword{methods}
