%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLanguage.RspLanguage}
\alias{getLanguage.RspLanguage}
\alias{RspLanguage.getLanguage}
\alias{getLanguage.RspLanguage}
\alias{getLanguage,RspLanguage-method}


\title{Gets the language string}

\description{
 Gets the language string.
}

\usage{\method{getLanguage}{RspLanguage}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
