%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getMetadata.RspProduct}
\alias{getMetadata.RspProduct}
\alias{RspProduct.getMetadata}
\alias{getMetadata,RspProduct-method}

\title{Gets the metadata an RSP product}

\description{
 Gets the metadata an RSP product.
}

\usage{
\method{getMetadata}{RspProduct}(object, name=NULL, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspProduct}}.
}
\keyword{internal}
\keyword{methods}
