%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.FileRspResponse}
\alias{getName.FileRspResponse}
\alias{FileRspResponse.getName}
\alias{getName,FileRspResponse-method}

\title{Gets the (base)name of the current RSP file}

\description{
 Gets the (base)name of the current RSP file.
}

\usage{
\method{getName}{FileRspResponse}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{FileRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
