%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspPage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getName.RspPage}
\alias{getName.RspPage}
\alias{RspPage.getName}
\alias{getName.RspPage}
\alias{getName,RspPage-method}


\title{Gets the (base)name of the current RSP file}

\description{
 Gets the (base)name of the current RSP file.
}

\usage{\method{getName}{RspPage}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspPage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
