%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getNewline.RspLanguage}
\alias{getNewline.RspLanguage}
\alias{RspLanguage.getNewline}
\alias{getNewline.RspLanguage}
\alias{getNewline,RspLanguage-method}


\title{Gets the newline string specific for a given RSP response language}

\description{
 Gets the newline string specific for a given RSP response language.
}

\usage{\method{getNewline}{RspLanguage}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
