%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getOutput.FileRspResponse}
\alias{getOutput.FileRspResponse}
\alias{FileRspResponse.getOutput}
\alias{getOutput,FileRspResponse-method}

\title{Gets the output for an RSP response}

\description{
 Gets the output for an RSP response.
}

\usage{
\method{getOutput}{FileRspResponse}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base:connections]{connection}} or a filename.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{FileRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
