%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspPage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.RspPage}
\alias{getPath.RspPage}
\alias{RspPage.getPath}
\alias{getPath,RspPage-method}

\title{Gets the path of the directory of the current RSP file}

\description{
 Gets the path of the directory of the current RSP file.
}

\usage{
\method{getPath}{RspPage}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspPage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
