%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPort.HttpDaemon}
\alias{getPort.HttpDaemon}
\alias{HttpDaemon.getPort}
\alias{getPort.HttpDaemon}
\alias{getPort,HttpDaemon-method}


\title{Gets the socket port of the HTTP daemon}

\description{
 Gets the socket port of the HTTP daemon, if started.
}

\synopsis{getPort.HttpDaemon(static, ...)}
\usage{HttpDaemon$getPort(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}} if started, otherwise \code{\link[base]{NA}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
