%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getProtocol.HttpRequest}
\alias{getProtocol.HttpRequest}
\alias{HttpRequest.getProtocol}
\alias{getProtocol.HttpRequest}
\alias{getProtocol,HttpRequest-method}


\title{Gets the name and version of the protocol used to make this request}

\description{
 Gets the name and version of the protocol used to make this request, e.g. HTTP/1.1.
}

\usage{\method{getProtocol}{HttpRequest}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{getScheme}}().
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
