%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRealPath.HttpRequest}
\alias{getRealPath.HttpRequest}
\alias{HttpRequest.getRealPath}
\alias{getRealPath.HttpRequest}
\alias{getRealPath,HttpRequest-method}


\title{Gets the file system path for a given URI}

\description{
 Gets the file system path for a given URI.
}

\usage{\method{getRealPath}{HttpRequest}(this, uri, ...)}

\arguments{
  \item{uri}{A URI as a \code{\link[base]{character}} string.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
