%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRemoteAddress.HttpRequest}
\alias{getRemoteAddress.HttpRequest}
\alias{HttpRequest.getRemoteAddress}
\alias{getRemoteAddress.HttpRequest}
\alias{getRemoteAddress,HttpRequest-method}


\title{Gets the IP address of the client that sent the request}

\description{
 Gets the IP address of the client that sent the request.
}

\usage{\method{getRemoteAddress}{HttpRequest}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{getRemoteHost}}().
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
