%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getReturn.RspCodeChunk}
\alias{getReturn.RspCodeChunk}
\alias{RspCodeChunk.getReturn}
\alias{getReturn.RspCodeChunk}
\alias{getReturn,RspCodeChunk-method}


\title{Checks whether the value of the evaluated code chunk should be returned or not}

\description{
 Checks whether the value of the evaluated code chunk should be returned or not.
}

\usage{
\method{getReturn}{RspCodeChunk}(code, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspCodeChunk}}.
}
\keyword{internal}
\keyword{methods}
