%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getServerPort.HttpRequest}
\alias{getServerPort.HttpRequest}
\alias{HttpRequest.getServerPort}
\alias{getServerPort.HttpRequest}
\alias{getServerPort,HttpRequest-method}


\title{Gets the port number on which this request was received}

\description{
 Gets the port number on which this request was received. 
}

\usage{\method{getServerPort}{HttpRequest}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{getServerPort}}().
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
