%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSuffixSpecs.RspConstruct}
\alias{getSuffixSpecs.RspConstruct}
\alias{RspConstruct.getSuffixSpecs}
\alias{getSuffixSpecs,RspConstruct-method}

\title{Gets the suffix specifications}

\description{
 Gets the suffix specifications.
}

\usage{
\method{getSuffixSpecs}{RspConstruct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a trimmed \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
