%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspExpression.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSuffixSpecs.RspExpression}
\alias{getSuffixSpecs.RspExpression}
\alias{RspExpression.getSuffixSpecs}
\alias{getSuffixSpecs.RspExpression}
\alias{getSuffixSpecs,RspExpression-method}


\title{Gets the suffix specifications}

\description{
 Gets the suffix specifications.
}

\usage{\method{getSuffixSpecs}{RspExpression}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a trimmed \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspExpression}}.
}
\keyword{internal}
\keyword{methods}
