%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getText.RspEvalDirective}
\alias{getText.RspEvalDirective}
\alias{RspEvalDirective.getText}
\alias{getText.RspEvalDirective}
\alias{getText,RspEvalDirective-method}


\title{Gets the text}

\description{
 Gets the text.
}

\usage{\method{getText}{RspEvalDirective}(directive, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspEvalDirective}}.
}
\keyword{internal}
\keyword{methods}
