%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspArtifact}
\alias{getType.RspArtifact}
\alias{RspArtifact.getType}
\alias{getType.RspArtifact}
\alias{getType,RspArtifact-method}


\title{Gets the type of an RSP artifact}

\description{
 Gets the type of an RSP artifact.
}

\usage{\method{getType}{RspArtifact}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspArtifact}}.
}
\keyword{internal}
\keyword{methods}
