%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspDocument}
\alias{getType.RspDocument}
\alias{RspDocument.getType}
\alias{getType,RspDocument-method}

\title{Gets the type of the RspDocument}

\description{
 Gets the type of the RspDocument.
}

\usage{
\method{getType}{RspDocument}(object, default=NA, as=c("text", "IMT"), ...)
}

\arguments{
  \item{default}{If unknown/not set, the default content type to return.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
