%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspFileArtifact}
\alias{getType.RspFileArtifact}
\alias{RspFileArtifact.getType}
\alias{getType.RspFileArtifact}
\alias{getType,RspFileArtifact-method}


\title{Gets the type of an RSP file artifact}

\description{
 Gets the type of an RSP file artifact.
}

\usage{\method{getType}{RspFileArtifact}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspFileArtifact}}.
}
\keyword{internal}
\keyword{methods}
