%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.SourceCode}
\alias{getType.SourceCode}
\alias{SourceCode.getType}
\alias{getType.SourceCode}
\alias{getType,SourceCode-method}


\title{Gets the type of the SourceCode}

\description{
 Gets the type of the SourceCode.
}

\usage{\method{getType}{SourceCode}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SourceCode}}.
}
\keyword{internal}
\keyword{methods}
