%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlRspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getVerbatim.HtmlRspLanguage}
\alias{getVerbatim.HtmlRspLanguage}
\alias{HtmlRspLanguage.getVerbatim}
\alias{getVerbatim.HtmlRspLanguage}
\alias{getVerbatim,HtmlRspLanguage-method}


\title{Gets a verbatim string specifically for the HTML language}

\description{
 Gets a verbatim string specifically for the HTML language.
}

\usage{\method{getVerbatim}{HtmlRspLanguage}(object, ..., newline=NULL)}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HtmlRspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
