%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getVerbatim.RspIncludeDirective}
\alias{getVerbatim.RspIncludeDirective}
\alias{RspIncludeDirective.getVerbatim}
\alias{getVerbatim,RspIncludeDirective-method}

\title{Checks if verbatim include should be used or not}

\description{
 Checks if verbatim include should be used or not.
}

\usage{
\method{getVerbatim}{RspIncludeDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{logical}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspIncludeDirective}}.
}
\keyword{internal}
\keyword{methods}
