%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getVerbatim.RspLanguage}
\alias{getVerbatim.RspLanguage}
\alias{RspLanguage.getVerbatim}
\alias{getVerbatim.RspLanguage}
\alias{getVerbatim,RspLanguage-method}


\title{Gets a verbatim string specific for a given RSP response language}

\description{
 Gets a verbatim string specific for a given RSP response language.
}

\usage{\method{getVerbatim}{RspLanguage}(this, ...)}

\arguments{
  \item{...}{R objects to be pasted together.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspLanguage}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
