%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getWrap.RspIncludeDirective}
\alias{getWrap.RspIncludeDirective}
\alias{RspIncludeDirective.getWrap}
\alias{getWrap,RspIncludeDirective-method}

\title{Get the wrap length}

\description{
 Get the wrap length.
}

\usage{
\method{getWrap}{RspIncludeDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}, or \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspIncludeDirective}}.
}
\keyword{internal}
\keyword{methods}
