%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasParameter.HttpRequest}
\alias{hasParameter.HttpRequest}
\alias{HttpRequest.hasParameter}
\alias{hasParameter.HttpRequest}
\alias{hasParameter,HttpRequest-method}


\title{Checks if a parameter exists}

\description{
 Checks if a parameter exists.
}

\usage{\method{hasParameter}{HttpRequest}(this, name, ...)}

\arguments{
  \item{name}{Name of parameter to be checked.}
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the parameter exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.rsp:getParameter.HttpRequest]{*getParameter}()}.
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
