%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasProcessor.RspArtifact}
\alias{hasProcessor.RspArtifact}
\alias{RspArtifact.hasProcessor}
\alias{hasProcessor.RspArtifact}
\alias{hasProcessor,RspArtifact-method}


\title{Checks whether a postprocessor exist or not for an RSP artifact}

\description{
 Checks whether a postprocessor exist or not for an RSP artifact.
}

\usage{\method{hasProcessor}{RspArtifact}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if one exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
