%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasProcessor.RspProduct}
\alias{hasProcessor.RspProduct}
\alias{RspProduct.hasProcessor}
\alias{hasProcessor,RspProduct-method}

\title{Checks whether a processor exist or not for an RSP product}

\description{
 Checks whether a processor exist or not for an RSP product.
}

\usage{
\method{hasProcessor}{RspProduct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns \code{\link[base:logical]{TRUE}} if one exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
