%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HtmlRspLanguage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{HtmlRspLanguage}
\docType{class}
\alias{HtmlRspLanguage}


 \title{The HtmlRspLanguage class}

 \description{
  Package:  R.rsp \cr
\bold{Class HtmlRspLanguage}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.rsp]{RspLanguage}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{HtmlRspLanguage}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{HtmlRspLanguage}\cr
extends \link[R.rsp]{RspLanguage}\cr


 }

 \usage{HtmlRspLanguage(...)}

 \arguments{
   \item{...}{Arguments passed to the constructor of the \code{\link{RspLanguage}}.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:escape.HtmlRspLanguage]{escape}} \tab Escapes a string specifically for the HTML language.\cr
 \tab \code{\link[R.rsp:getComment.HtmlRspLanguage]{getComment}} \tab Gets a comment string specifically for the HTML language.\cr
 \tab \code{\link[R.rsp:getVerbatim.HtmlRspLanguage]{getVerbatim}} \tab Gets a verbatim string specifically for the HTML language.\cr
}


 \bold{Methods inherited from RspLanguage}:\cr
escape, getComment, getLanguage, getNewline, getVerbatim

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save

 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
