%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpRequest}
\docType{class}
\alias{HttpRequest}


\title{The HttpRequest class}

\description{
 Package:  R.rsp \cr
\bold{Class HttpRequest}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{HttpRequest}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{HttpRequest}\cr
extends \link[R.oo]{Object}\cr


}

\usage{
HttpRequest(requestUri=NULL, parameters=list(), ...)
}

\arguments{
  \item{requestUri}{A \code{\link[base]{character}} string of the requested URI.}
  \item{parameters}{A named \code{\link[base]{list}} of parameter values.}
  \item{...}{Not used.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{as.character} \tab  -\cr
 \tab \code{getContentLength} \tab  -\cr
 \tab \code{getContentType} \tab  -\cr
 \tab \code{getContextPath} \tab  -\cr
 \tab \code{getDateHeader} \tab  -\cr
 \tab \code{getHeader} \tab  -\cr
 \tab \code{getParameter} \tab  -\cr
 \tab \code{getParameters} \tab  -\cr
 \tab \code{getProtocol} \tab  -\cr
 \tab \code{getQueryString} \tab  -\cr
 \tab \code{getRealPath} \tab  -\cr
 \tab \code{getRemoteAddress} \tab  -\cr
 \tab \code{getRemoteHost} \tab  -\cr
 \tab \code{getRemoteUser} \tab  -\cr
 \tab \code{getRequestUri} \tab  -\cr
 \tab \code{getRequestUrl} \tab  -\cr
 \tab \code{getScheme} \tab  -\cr
 \tab \code{getServerName} \tab  -\cr
 \tab \code{getServerPort} \tab  -\cr
 \tab \code{getServletPath} \tab  -\cr
 \tab \code{hasParameter} \tab  -\cr
 \tab \code{nbrOfParameters} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


}

\author{Henrik Bengtsson}

\keyword{classes}
\keyword{internal}
