%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  importRsp.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{importRsp}
\alias{importRsp.default}
\alias{importRsp}

\title{Imports an RSP file}

\description{
 Imports an RSP file.
}

\usage{\method{importRsp}{default}(...)}

\arguments{
  \item{...}{Arguments passed to \code{\link{sourceRsp}}().}
}

\value{
  Returns the compile output of an RSP template as a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link{sourceRsp}}().
}




\keyword{file}
\keyword{IO}
\keyword{internal}
