%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isStarted.HttpDaemon}
\alias{isStarted.HttpDaemon}
\alias{HttpDaemon.isStarted}
\alias{isStarted.HttpDaemon}
\alias{isStarted,HttpDaemon-method}


\title{Checks if the HTTP daemon is started}

\description{
 Checks if the HTTP daemon is started.
}

\synopsis{isStarted.HttpDaemon(x, ...)}
\usage{HttpDaemon$isStarted(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the server is started, otherwise \code{\link[base:logical]{FALSE}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.rsp:start.HttpDaemon]{*start}()} and \code{\link[R.rsp:stop.HttpDaemon]{*stop}()}.
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
