%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{makeSourceCode.RspSourceCodeFactory}
\alias{makeSourceCode.RspSourceCodeFactory}
\alias{RspSourceCodeFactory.makeSourceCode}
\alias{makeSourceCode.RspSourceCodeFactory}
\alias{makeSourceCode,RspSourceCodeFactory-method}


\title{Makes a SourceCode object}

\description{
 Makes a SourceCode object.
}

\usage{\method{makeSourceCode}{RspSourceCodeFactory}(this, ...)}

\arguments{
  \item{...}{Arguments passed to the language-specific
             \code{\link{SourceCode}} constructor.}
}

\value{
 Returns a \code{\link{SourceCode}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
