%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{mergeTexts.RspDocument}
\alias{mergeTexts.RspDocument}
\alias{RspDocument.mergeTexts}
\alias{mergeTexts.RspDocument}
\alias{mergeTexts,RspDocument-method}


\title{Merge neighboring RspText:s}

\description{
 Merge neighboring RspText:s by pasting them together.
}

\usage{\method{mergeTexts}{RspDocument}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link{RspDocument}} with equal or fever number of
  \code{\link{RspExpression}}:s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
