%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfLines.RspString}
\alias{nbrOfLines.RspString}
\alias{RspString.nbrOfLines}
\alias{nbrOfLines,RspString-method}

\title{Gets the number of lines in an RSP string}

\description{
 Gets the number of lines in an RSP string.
}

\usage{
\method{nbrOfLines}{RspString}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a non-negative \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspString}}.
}
\keyword{internal}
\keyword{methods}
