%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpRequest.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfParameters.HttpRequest}
\alias{nbrOfParameters.HttpRequest}
\alias{HttpRequest.nbrOfParameters}
\alias{nbrOfParameters.HttpRequest}
\alias{nbrOfParameters,HttpRequest-method}


\title{Gets the number of parameters}

\description{
 Gets the number of parameters.
}

\usage{\method{nbrOfParameters}{HttpRequest}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.rsp:getParameter.HttpRequest]{*getParameter}()}.
  \code{\link[R.rsp:hasParameter.HttpRequest]{*hasParameter}()}.
  For more information see \code{\link{HttpRequest}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
