%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% The HttpDaemon class
\alias{appendRootPaths}
\alias{finalize.HttpDaemon}
\alias{getConfig}
\alias{getCount}
\alias{getCount.HttpDaemon}
\alias{setCount}
\alias{setCount.HttpDaemon}
\alias{getDefaultFilename}
\alias{getHttpRequest}
\alias{getPort}
\alias{getRootPath}
\alias{getRootPaths}
\alias{insertRootPaths}
\alias{isStarted}
\alias{openUrl}
\alias{processRsp}
\alias{restart}
\alias{restart.default}
\alias{setRootPaths}
\alias{sourceTcl}
\alias{startHelp}
\alias{stop}
\alias{writeResponse}

% The HttpDaemonResponse class
\alias{flush}

% The HttpRequest class
\alias{getContentLength}
\alias{getContentType}
\alias{getContextPath}
\alias{getContextPath.HttpRequest}
\alias{getDateHeader}
\alias{getDateHeader.HttpRequest}
\alias{getHeader}
\alias{getHeader.HttpRequest}
\alias{getParameter}
\alias{getParameters}
\alias{getProtocol}
\alias{getQueryString}
\alias{getQueryString.HttpRequest}
\alias{getRealPath}
\alias{getRemoteAddress}
\alias{getRemoteHost}
\alias{getRemoteUser}
\alias{getRemoteUser.HttpRequest}
\alias{getRequestUri}
\alias{getRequestUri.HttpRequest}
\alias{getRequestUrl}
\alias{getRequestUrl.HttpRequest}
\alias{getScheme}
\alias{getServerName}
\alias{getServerPort}
\alias{getServletPath}
\alias{getServletPath.HttpRequest}
\alias{hasParameter}
\alias{nbrOfParameters}

% The RspResponse class
\alias{getOutput}
\alias{import}
\alias{write}

% The RspLanguage class
\alias{escape}
\alias{getComment}
\alias{getLanguage}
\alias{getNewline}
\alias{getVerbatim}

% The HtmlRspLanguage class

% The RspPage class

% Miscellanoues stuff
\alias{hexToInt}
\alias{includeRsp.default}
\alias{stop.default}
\alias{write.default}
\alias{urlDecode}
\alias{compileRsp}
\alias{compileRsp.default}
\alias{epsDev}
\alias{dvips}
\alias{ps2pdf}
\alias{rspCapture}
\alias{evalWithEcho}
\alias{sourceWithTrim}
\alias{sourceWithTrim.default}
\alias{sourceRspV2}
\alias{sourceRspV2.default}

\description{
This page contains aliases for all "non-documented" objects that
\code{R CMD check} detects in this package.

Almost all of them are \emph{generic} functions that have specific
document for the corresponding method coupled to a specific class.
Other functions are re-defined by \code{setMethodS3()} to
\emph{default} methods. Neither of these two classes are non-documented
in reality.
The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
