%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{parse.RSourceCode}
\alias{parse.RSourceCode}
\alias{RSourceCode.parse}
\alias{parse.RSourceCode}
\alias{parse,RSourceCode-method}


\title{Parses the R code}

\description{
 Parses the R code.
}

\usage{\method{parse}{RSourceCode}(this, ...)}

\arguments{
  \item{...}{Optional arguments passed to \code{\link[base]{parse}}().}
}

\value{
 Returns an \code{\link[base]{expression}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RSourceCode}}.
}
\keyword{internal}
\keyword{methods}
