%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{parse.RspRSourceCode}
\alias{parse.RspRSourceCode}
\alias{RspRSourceCode.parse}
\alias{parse,RspRSourceCode-method}

\title{Parses the R code}

\description{
 Parses the R code.
}

\usage{
\method{parse}{RspRSourceCode}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link[base]{expression}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspRSourceCode}}.
}
\keyword{internal}
\keyword{methods}
