%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{parse.SourceCode}
\alias{parse.SourceCode}
\alias{SourceCode.parse}
\alias{parse.SourceCode}
\alias{parse,SourceCode-method}


\title{Parses the source code}

\description{
 Parses the source code.
}

\usage{\method{parse}{SourceCode}(...)}

\arguments{
  \item{...}{Optional arguments passed to language parse.}
}

\value{
 Returns an \code{\link[base]{expression}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SourceCode}}.
}
\keyword{internal}
\keyword{methods}
