%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{parseDirective.RspUnparsedDirective}
\alias{parseDirective.RspUnparsedDirective}
\alias{RspUnparsedDirective.parseDirective}
\alias{parseDirective,RspUnparsedDirective-method}

\title{Parses the unknown RSP directive for its class}

\description{
 Parses the unknown RSP directive for its class.
}

\usage{
\method{parseDirective}{RspUnparsedDirective}(expr, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspDirective}} of known class.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspUnparsedDirective}}.
}
\keyword{internal}
\keyword{methods}
