%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspExpression.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{parseExpression.RspUnparsedExpression}
\alias{parseExpression.RspUnparsedExpression}
\alias{RspUnparsedExpression.parseExpression}
\alias{parseExpression,RspUnparsedExpression-method}

\title{Parses the unknown RSP expression for its class}

\description{
 Parses the unknown RSP expression for its class.
}

\usage{
\method{parseExpression}{RspUnparsedExpression}(expr, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspExpression}} of known class.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspUnparsedExpression}}.
}
\keyword{internal}
\keyword{methods}
