%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{postprocess.RspArtifact}
\alias{postprocess.RspArtifact}
\alias{RspArtifact.postprocess}
\alias{postprocess.RspArtifact}
\alias{postprocess,RspArtifact-method}


\title{Postprocesses an RSP artifact}

\description{
 Postprocesses an RSP artifact.
}

\usage{\method{postprocess}{RspArtifact}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
