%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  rargs.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{print.RspArguments}
\alias{print.RspArguments}
\alias{RspArguments.print}
\alias{print,RspArguments-method}

\title{Prints RSP arguments}

\description{
 Prints RSP arguments returned by \code{\link{rargs}}().
}

\usage{
\method{print}{RspArguments}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link{rargs}}().
}
\keyword{internal}
\keyword{methods}
