%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspArtifact}
\alias{print.RspArtifact}
\alias{RspArtifact.print}
\alias{print.RspArtifact}
\alias{print,RspArtifact-method}


\title{Prints a summary of an RSP artifact}

\description{
 Prints a summary of an RSP artifact.
}

\usage{\method{print}{RspArtifact}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspArtifact}}.
}
\keyword{internal}
\keyword{methods}
