%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspArtifact.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspFileArtifact}
\alias{print.RspFileArtifact}
\alias{RspFileArtifact.print}
\alias{print.RspFileArtifact}
\alias{print,RspFileArtifact-method}


\title{Prints a summary of an RSP file artifact}

\description{
 Prints a summary of an RSP file artifact.
}

\usage{\method{print}{RspFileArtifact}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspFileArtifact}}.
}
\keyword{internal}
\keyword{methods}
