%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspFileProduct}
\alias{print.RspFileProduct}
\alias{RspFileProduct.print}
\alias{print.RspFileProduct}
\alias{print,RspFileProduct-method}


\title{Prints a summary of an RSP file product}

\description{
 Prints a summary of an RSP file product.
}

\usage{\method{print}{RspFileProduct}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspFileProduct}}.
}
\keyword{internal}
\keyword{methods}
