%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspObject.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspObject}
\alias{print.RspObject}
\alias{RspObject.print}
\alias{print,RspObject-method}
\alias{print.RspDocument}
\alias{print.RspFileProduct}
\alias{print.RspProduct}
\alias{print.RspSourceCode}
\alias{print.RspString}
\alias{print.RspStringProduct}

\title{Prints a summary of an RSP object}

\description{
 Prints a summary of an RSP object.
}

\usage{
\method{print}{RspObject}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspObject}}.
}
\keyword{internal}
\keyword{methods}
