%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspProduct}
\alias{print.RspProduct}
\alias{RspProduct.print}
\alias{print.RspProduct}
\alias{print,RspProduct-method}


\title{Prints a summary of an RSP product}

\description{
 Prints a summary of an RSP product.
}

\usage{
\method{print}{RspProduct}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspProduct}}.
}
\keyword{internal}
\keyword{methods}
