%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspSourceCode}
\alias{print.RspSourceCode}
\alias{RspSourceCode.print}
\alias{print.RspSourceCode}
\alias{print,RspSourceCode-method}


\title{Prints the source code}

\description{
 Prints the source code.
}

\usage{\method{print}{RspSourceCode}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspSourceCode}}.
}
\keyword{internal}
\keyword{methods}
