%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspStringProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.RspStringProduct}
\alias{print.RspStringProduct}
\alias{RspStringProduct.print}
\alias{print.RspStringProduct}
\alias{print,RspStringProduct-method}


\title{Prints an RSP string product}

\description{
 Prints an RSP string product.
}

\usage{
\method{print}{RspStringProduct}(x, ...)
}

\arguments{
  \item{...}{Additional arguments passed to \code{\link[base]{print}}().}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspStringProduct}}.
}
\keyword{internal}
\keyword{methods}
