%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.SourceCode}
\alias{print.SourceCode}
\alias{SourceCode.print}
\alias{print.SourceCode}
\alias{print,SourceCode-method}


\title{Prints the source code}

\description{
 Prints the source code.
}

\usage{\method{print}{SourceCode}(x, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SourceCode}}.
}
\keyword{internal}
\keyword{methods}
