%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.RspFileProduct}
\alias{process.RspFileProduct}
\alias{RspFileProduct.process}
\alias{process.RspFileProduct}
\alias{process,RspFileProduct-method}


\title{Processes an RSP file product}

\description{
 Processes an RSP file product.
}

\usage{\method{process}{RspFileProduct}(object, type=NULL, workdir=".", ..., verbose=FALSE)}

\arguments{
  \item{type}{A \code{\link[base]{character}} string specifying the content type.}
  \item{workdir}{The working directory.}
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
  Returns the processed RSP product, e.g. the pathname to a PDF file.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
