%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{process.RspProduct}
\alias{process.RspProduct}
\alias{RspProduct.process}
\alias{process.RspProduct}
\alias{process,RspProduct-method}


\title{Processes an RSP product}

\description{
 Processes an RSP product.
}

\usage{\method{process}{RspProduct}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns the processed RSP product, which is another \code{\link{RspProduct}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
