%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{requireAttributes.RspDirective}
\alias{requireAttributes.RspDirective}
\alias{RspDirective.requireAttributes}
\alias{requireAttributes,RspDirective-method}

\title{Asserts that certain attributes exist}

\description{
 Asserts that certain attributes exist.
}

\usage{
\method{requireAttributes}{RspDirective}(this, names, condition=c("all", "any"), ...)
}

\arguments{
  \item{condition}{A \code{\link[base]{character}} specifying the condition to be tested.}
  \item{...}{Not used.}
}

\value{
 Returns itself (invisibly).
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDirective}}.
}
\keyword{internal}
\keyword{methods}
