%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$restart}
\alias{HttpDaemon$restart}
\alias{restart.HttpDaemon}
\alias{HttpDaemon.restart}
\alias{restart,HttpDaemon-method}

\title{Restarts the HTTP daemon}

\description{
 Restarts the HTTP daemon.
}

\usage{
## Static method (use this):
## HttpDaemon$restart(...)

## Don't use the below:
\method{restart}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.rsp:isStarted.HttpDaemon]{*isStarted}()}.
  \code{\link[R.rsp:start.HttpDaemon]{*start}()}.
  \code{\link[R.rsp:stop.HttpDaemon]{*stop}()}.
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
