%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.RRspPackage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RRspPackage}
\docType{class}
\alias{RRspPackage}


\title{The RRspPackage class}

\description{
 Package:  R.rsp \cr
\bold{Class RRspPackage}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.oo]{Package}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{RRspPackage}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{RRspPackage}\cr
extends \link[R.oo]{Package}\cr


}

\usage{
RRspPackage(...)
}

\arguments{
  \item{...}{Not used.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:capabilitiesOf.RRspPackage]{capabilitiesOf}} \tab Checks which tools are supported.\cr
 \tab \code{isCapableOf} \tab  -\cr
}


 \bold{Methods inherited from Package}:\cr
as.character, getAuthor, getBundle, getBundlePackages, getChangeLog, getClasses, getContents, getDataPath, getDate, getDescription, getDescriptionFile, getDocPath, getEnvironment, getExamplePath, getHistory, getHowToCite, getLicense, getMaintainer, getName, getNews, getPath, getPosition, getTitle, getUrl, getVersion, isLoaded, isOlderThan, ll, load, showChangeLog, showContents, showDescriptionFile, showHistory, showHowToCite, showNews, startupMessage, unload

 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


}

\author{Henrik Bengtsson}


\keyword{classes}
\keyword{internal}
