%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspParser.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RspParser}
\docType{class}
\alias{RspParser}


\title{The RspParser class}

\description{
 Package:  R.rsp \cr
\bold{Class RspParser}\cr

\code{logical}\cr
\code{~~|}\cr
\code{~~+--}\code{RspParser}\cr

\bold{Directly known subclasses:}\cr
\cr

public class \bold{RspParser}\cr
extends logical\cr



 An RspParser is parser for the RSP language.
}

\usage{
RspParser(...)
}

\arguments{
  \item{...}{Not used.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:parseDocument.RspParser]{parseDocument}} \tab Parse an RSP string into and RSP document.\cr
 \tab \code{\link[R.rsp:parseRaw.RspParser]{parseRaw}} \tab Parses the string into blocks of text and RSP.\cr
}


 \bold{Methods inherited from logical}:\cr
Ops,nonStructure,vector-method, Ops,structure,vector-method, Ops,vector,nonStructure-method, Ops,vector,structure-method, as.data.frame, as.raster, coerce,ANY,logical-method, whichVector


}

\author{Henrik Bengtsson}


\keyword{classes}
\keyword{internal}
