%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RspProduct}
\docType{class}
\alias{RspProduct}


\title{The RspProduct class}

\description{
 Package:  R.rsp \cr
\bold{Class RspProduct}\cr

\code{logical}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.rsp]{RspObject}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\emph{\code{RspProduct}}\cr

\bold{Directly known subclasses:}\cr
\link[R.rsp]{RspFileProduct}, \emph{\link[R.rsp]{RspRSourceCode}}, \emph{\link[R.rsp]{RspShSourceCode}}, \emph{\link[R.rsp]{RspSourceCode}}, \emph{\link[R.rsp]{RspStringProduct}}\cr

public abstract class \bold{RspProduct}\cr
extends \link[R.rsp]{RspObject}\cr



 An RspProduct object represents an RSP product generated by processing
 an RSP document.
}

\usage{
RspProduct(object=NA, ...)
}

\arguments{
  \item{object}{The RSP product.}
  \item{...}{Arguments passed to \code{\link{RspObject}}.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{!} \tab  -\cr
 \tab \code{findProcessor} \tab  -\cr
 \tab \code{getMetadata} \tab  -\cr
 \tab \code{getType} \tab  -\cr
 \tab \code{hasProcessor} \tab  -\cr
 \tab \code{print} \tab  -\cr
 \tab \code{process} \tab  -\cr
 \tab \code{setMetadata} \tab  -\cr
 \tab \code{view} \tab  -\cr
}


 \bold{Methods inherited from RspObject}:\cr
getAttribute, getAttributes, hasAttribute, print, setAttribute, setAttributes

 \bold{Methods inherited from logical}:\cr
as.data.frame, as.raster, whichVector


}

\author{Henrik Bengtsson}


\keyword{classes}
\keyword{internal}
