%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RspRSourceCodeFactory}
\docType{class}
\alias{RspRSourceCodeFactory}


\title{The RspRSourceCodeFactory class}

\description{
 Package:  R.rsp \cr
\bold{Class RspRSourceCodeFactory}\cr

\code{character}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.rsp]{RspSourceCodeFactory}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\code{RspRSourceCodeFactory}\cr

\bold{Directly known subclasses:}\cr
\cr

public class \bold{RspRSourceCodeFactory}\cr
extends \emph{\link[R.rsp]{RspSourceCodeFactory}}\cr



 An RspRSourceCodeFactory is an \code{\link{RspSourceCodeFactory}} for
 the R language.
}

\usage{
RspRSourceCodeFactory(...)
}

\arguments{
  \item{...}{Not used.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{exprToCode} \tab  -\cr
 \tab \code{getCompleteCode} \tab  -\cr
}


 \bold{Methods inherited from RspSourceCodeFactory}:\cr
exprToCode, getCompleteCode, getLanguage, makeSourceCode, toSourceCode

 \bold{Methods inherited from character}:\cr
all.equal, as.Date, as.POSIXlt, as.data.frame, as.raster, downloadFile, formula, getDLLRegisteredRoutines, isOpen, toAsciiRegExprPattern, toFileListTree, toLatex, uses


}

\author{Henrik Bengtsson}


\keyword{classes}
\keyword{internal}
