%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$sourceTcl}
\alias{HttpDaemon$sourceTcl}
\alias{sourceTcl.HttpDaemon}
\alias{HttpDaemon.sourceTcl}
\alias{sourceTcl,HttpDaemon-method}

\title{Loads the Tcl source for the HTTP daemon into R}

\description{
 Loads the Tcl source for the HTTP daemon into R.
}

\usage{
## Static method (use this):
## HttpDaemon$sourceTcl(...)

## Don't use the below:
\method{sourceTcl}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
