%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  splitUrl.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{splitUrl}
\alias{splitUrl.default}
\alias{splitUrl}

\title{Decomposes a URL into its components}

\description{
 Decomposes a URL into its components.
}

\usage{
\method{splitUrl}{default}(url, ...)
}

\arguments{
  \item{url}{A \code{\link[base]{character}} string.}
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}} of URL components.
}

\author{Henrik Bengtsson}

\references{
 [1] \url{https://en.wikipedia.org/wiki/URI_scheme}
}




\keyword{IO}
\keyword{programming}
\keyword{internal}
