%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{startHelp.HttpDaemon}
\alias{startHelp.HttpDaemon}
\alias{HttpDaemon.startHelp}
\alias{startHelp.HttpDaemon}
\alias{startHelp,HttpDaemon-method}


\title{Starts the HTTP daemon and launches the help page}

\description{
 Starts the HTTP daemon and launches the help page.
}

\usage{\method{startHelp}{HttpDaemon}(static, ...)}

\arguments{
  \item{...}{Arguments passed to \code{\link[R.rsp:openUrl.HttpDaemon]{*openUrl}()}.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
