%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{subset.RspDocument}
\alias{subset.RspDocument}
\alias{RspDocument.subset}
\alias{subset,RspDocument-method}

\title{Subsets an RspDocument}

\description{
 Subsets an RspDocument.
}

\usage{
\method{subset}{RspDocument}(x, subset, ...)
}

\arguments{
  \item{subset}{An \code{\link[base]{expression}} used for subsetting.}
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspDocument}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
