%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{tangle.RSourceCode}
\alias{tangle.RSourceCode}
\alias{RSourceCode.tangle}
\alias{tangle.RSourceCode}
\alias{tangle,RSourceCode-method}


\title{Drops all text-outputting calls from the R code}

\description{
 Drops all text-outputting calls from the R code.
}

\usage{\method{tangle}{RSourceCode}(code, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RSourceCode}} objects.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RSourceCode}}.
}
\keyword{internal}
\keyword{methods}
