%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspRSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{tangle.RspRSourceCode}
\alias{tangle.RspRSourceCode}
\alias{RspRSourceCode.tangle}
\alias{tangle,RspRSourceCode-method}

\title{Drops all text-outputting calls from the R code}

\description{
 Drops all text-outputting calls from the R code.
}

\usage{
\method{tangle}{RspRSourceCode}(code, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspRSourceCode}} objects without output calls.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspRSourceCode}}.
}
\keyword{internal}
\keyword{methods}
