%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{tangle.RspSourceCode}
\alias{tangle.RspSourceCode}
\alias{RspSourceCode.tangle}
\alias{tangle,RspSourceCode-method}
\alias{tangle.RspRSourceCode}

\title{Drops all text-outputting calls from the source code}

\description{
 Drops all text-outputting calls from the source code.
}

\usage{
\method{tangle}{RspSourceCode}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{RspSourceCode}} objects.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspSourceCode}}.
}
\keyword{internal}
\keyword{methods}
