%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{tangle.SourceCode}
\alias{tangle.SourceCode}
\alias{SourceCode.tangle}
\alias{tangle.SourceCode}
\alias{tangle,SourceCode-method}


\title{Drops all text-outputting calls from the source code}

\description{
 Drops all text-outputting calls from the source code.
}

\usage{\method{tangle}{SourceCode}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{SourceCode}} objects.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SourceCode}}.
}
\keyword{internal}
\keyword{methods}
