%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HttpDaemon$terminate}
\alias{HttpDaemon$terminate}
\alias{terminate.HttpDaemon}
\alias{HttpDaemon.terminate}
\alias{terminate,HttpDaemon-method}

\title{Terminates the HTTP daemon}

\description{
 Terminates the HTTP daemon.
}

\usage{
## Static method (use this):
## HttpDaemon$terminate(...)

## Don't use the below:
\method{terminate}{HttpDaemon}(static, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.rsp:isStarted.HttpDaemon]{*isStarted}()}.
  \code{\link[R.rsp:start.HttpDaemon]{*start}()}.
  \code{\link[R.rsp:restart.HttpDaemon]{*restart}()}.
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
