%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  toLatex.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{toLatex.character}
\alias{toLatex.character}
\alias{toLatex.default}
\alias{le}

\title{Escapes character strings to become LaTeX compatible}

\description{
 Escapes character strings to become LaTeX compatible.
}

\usage{
\method{toLatex}{character}(object, ...)
}

\arguments{
  \item{object}{A \code{\link[base]{character}} \code{\link[base]{vector}} of N strings.}
  \item{...}{Not used.}
}

\value{
  A \code{\link[base]{character}} \code{\link[base]{vector}} of N strings.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[utils]{toLatex}} for escaping \code{\link[utils]{sessionInfo}} objects.
}


\keyword{methods}
\keyword{internal}
