%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{trim.RspDocument}
\alias{trim.RspDocument}
\alias{RspDocument.trim}
\alias{trim,RspDocument-method}

\title{Trims each of the RSP constructs}

\description{
 Trims each of the RSP constructs.
}

\usage{
\method{trim}{RspDocument}(object, ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
 Returns the trimmed \code{\link{RspDocument}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
