%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{view.RspProduct}
\alias{view.RspProduct}
\alias{RspProduct.view}
\alias{view,RspProduct-method}
\alias{view.RspFileProduct}

\title{Views the RSP product}

\description{
 Views the RSP product.
}

\usage{
\method{view}{RspProduct}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns the RSP product (invisibly).
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspProduct}}.
}
\keyword{internal}
\keyword{methods}
