%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FileRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.FileRspResponse}
\alias{write.FileRspResponse}
\alias{FileRspResponse.write}
\alias{write.FileRspResponse}
\alias{write,FileRspResponse-method}


\title{Writes an RSP response to the predefined output file}

\description{
 Writes an RSP response to the predefined output file.
}

\usage{\method{write}{FileRspResponse}(this, ..., collapse="", sep="")}

\arguments{
  \item{...}{Objects to be pasted together and outputted.}
  \item{collapse}{A \code{\link[base]{character}} string to be used to collapse the objects.}
  \item{sep}{A \code{\link[base]{character}} string to separate the objects.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FileRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
