%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemonRspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{write.HttpDaemonRspResponse}
\alias{write.HttpDaemonRspResponse}
\alias{HttpDaemonRspResponse.write}
\alias{write.HttpDaemonRspResponse}
\alias{write,HttpDaemonRspResponse-method}


\title{Writes strings to an HttpDaemonRspResponse buffer}
\description{
 Writes strings to an HttpDaemonRspResponse buffer.
}

\usage{\method{write}{HttpDaemonRspResponse}(this, ..., collapse="", sep="")}

\arguments{
  \item{...}{One or several strings, which will be concatenated.}
  \item{sep}{The string to used for concatenating several strings.}
  \item{collapse}{The string to used collapse vectors together.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemonRspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
