%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspResponse.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeResponse.RspResponse}
\alias{writeResponse.RspResponse}
\alias{RspResponse.writeResponse}
\alias{writeResponse,RspResponse-method}
\alias{writeResponse.FileRspResponse}
\alias{writeResponse.HttpDaemonRspResponse}
\alias{RspResponse.write}
\alias{write.RspResponse}
\alias{write,RspResponse-method}


\title{Writes an RSP response to the predefined output}

\description{
 Writes an RSP response to the predefined output.
}

\usage{
\method{writeResponse}{RspResponse}(...)
}

\arguments{
  \item{...}{Objects to be pasted together and outputted.}
  \item{collapse}{A \code{\link[base]{character}} string to be used to collapse the objects.}
  \item{sep}{A \code{\link[base]{character}} string to separate the objects.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspResponse}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
