%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{asRspString.RspConstruct}
\alias{asRspString.RspConstruct}
\alias{RspConstruct.asRspString}
\alias{asRspString,RspConstruct-method}
\alias{asRspString.RspCode}
\alias{asRspString.RspCodeChunk}
\alias{asRspString.RspComment}
\alias{asRspString.RspDirective}
\alias{asRspString.RspDocument}
\alias{asRspString.RspText}
\alias{asRspString.RspUnParsedDirective}
\alias{asRspString.RspUnparsedDirective}

\title{Recreates an RSP string from an RspConstruct}

\description{
 Recreates an RSP string from an RspConstruct.
}

\usage{
\method{asRspString}{RspConstruct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns an \code{\link{RspString}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
