%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  buildNonSweaveVignette.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{buildNonSweaveVignette}
\alias{buildNonSweaveVignette}


\title{Builds a non-Sweave Rnw vignette}

\description{
 Builds a non-Sweave Rnw vignette.
}

\usage{
buildNonSweaveVignette(vign, envir=new.env(), ...)
}

\arguments{
  \item{pathname}{The vignette file to be built.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}




\keyword{file}
\keyword{IO}
\keyword{internal}
