%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  buildVignette.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{buildVignette}
\alias{buildVignette}


\title{Builds a vignette}

\description{
 Builds a vignette into a PDF or a HTML file.
}

\usage{
buildVignette(file, dir=".", latex=TRUE, tangle=TRUE, quiet=TRUE, clean=TRUE,
  engine=NULL, buildPkg=NULL, ...)
}

\arguments{
  \item{file}{A character string specifying the vignette source file.}
  \item{dir}{A character string specifying the working directory.}
  \item{latex}{A logical specifying whether a TeX produced by weave
    should be compiled by \link{texi2pdf} or not.}
  \item{tangle}{A logical specifying whether tangle should be run or not.}
  \item{quiet}{A logical specifying whether weave, \link{texi2pdf}, and
    tangle should be run in quiet mode or not.}
  \item{clean}{Remove all files generated by the build, even if there
    were copies there before.}
  \item{engine}{An optional \code{\link[base]{character}} string speciyfing the vignette
    engine to use for building the vignette.  If not specified, it will
    be inferred from the \code{\\VignetteEngine\{\}} markup in the
    vignette.  If that is not specified, the \code{"utils::Sweave"}
    engine will be used.}
  \item{buildPkg}{An optional \code{\link[base]{character}} string specifying the vignette
    package to use, if the vignette engine does not specify one.}
  \item{...}{Additional arguments passed to weave and tangle.}
}

\value{
  Returns the filename to the final vignette (PDF or HTML) product.
  If \code{latex = FALSE}, it may be a TeX file.
}

\author{Henrik Bengtsson}

\seealso{
  Since R devel (to become 3.1.0) r63076 (2013-06-27), there is
  \code{buildVignette()} in the \pkg{tools} package.
}





\keyword{documentation}
\keyword{file}
\keyword{IO}
\keyword{internal}
