%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{evaluate.RspSourceCode}
\alias{evaluate.RspSourceCode}
\alias{RspSourceCode.evaluate}
\alias{evaluate,RspSourceCode-method}

\title{Evaluates the source code}

\description{
 Evaluates the source code.
}

\usage{
\method{evaluate}{RspSourceCode}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns the last evaluated expression, iff any.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspSourceCode}}.
}
\keyword{internal}
\keyword{methods}
