%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findProcessor.RspProduct}
\alias{findProcessor.RspProduct}
\alias{RspProduct.findProcessor}
\alias{findProcessor,RspProduct-method}
\alias{findProcessor.RspFileProduct}

\title{Locates a processor for an RSP product}

\description{
 Locates a processor for an RSP product.
}

\usage{
\method{findProcessor}{RspProduct}(object, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{function}} that takes an \code{\link{RspProduct}} as input,
  or \code{\link[base]{NULL}} if no processor was found.
}

\author{Henrik Bengtsson}



\keyword{internal}
\keyword{methods}
\keyword{file}
\keyword{IO}
