%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspConstruct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAttributes.RspConstruct}
\alias{getAttributes.RspConstruct}
\alias{RspConstruct.getAttributes}
\alias{getAttributes,RspConstruct-method}
\alias{RspConstruct.getAttribute}
\alias{getAttribute.RspConstruct}
\alias{getAttribute,RspConstruct-method}


\title{Gets the attributes of an RSP expression}

\description{
 Gets the attributes of an RSP expression.
}

\usage{
\method{getAttributes}{RspConstruct}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a named \code{\link[base]{list}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspConstruct}}.
}
\keyword{internal}
\keyword{methods}
