%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getConfig.HttpDaemon}
\alias{getConfig.HttpDaemon}
\alias{HttpDaemon.getConfig}
\alias{getConfig.HttpDaemon}
\alias{getConfig,HttpDaemon-method}


\title{Retrieves the servers 'config' structure from Tcl}

\description{
 Retrieves the servers 'config' structure from Tcl.
}

\synopsis{getConfig.HttpDaemon(static, ...)}
\usage{HttpDaemon$getConfig(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a tclArray object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
