%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getHttpRequest.HttpDaemon}
\alias{getHttpRequest.HttpDaemon}
\alias{HttpDaemon.getHttpRequest}
\alias{getHttpRequest.HttpDaemon}
\alias{getHttpRequest,HttpDaemon-method}


\title{Gets the HTTP request}

\description{
 Gets the HTTP request.
}

\synopsis{getHttpRequest.HttpDaemon(static, ...)}
\usage{HttpDaemon$getHttpRequest(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link{HttpRequest}} object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
