%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getText.RspIncludeDirective}
\alias{getText.RspIncludeDirective}
\alias{RspIncludeDirective.getText}
\alias{getText,RspIncludeDirective-method}

\title{Gets the text}

\description{
 Gets the text.
}

\usage{
\method{getText}{RspIncludeDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspIncludeDirective}}.
}
\keyword{internal}
\keyword{methods}
