%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspText.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getText.RspText}
\alias{getText.RspText}
\alias{RspText.getText}
\alias{getText,RspText-method}

\title{Gets the text}

\description{
 Gets the text.
}

\usage{
\method{getText}{RspText}(text, unescape=FALSE, ...)
}

\arguments{
  \item{...}{Not used.}
  \item{unescaped}{If \code{\link[base:logical]{TRUE}}, character sequences \code{<\%\%} and
               \code{\%\%>} are unescaped to \code{<\%} and \code{\%>}.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspText}}.
}
\keyword{internal}
\keyword{methods}
