%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDirective.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTitle.RspPageDirective}
\alias{getTitle.RspPageDirective}
\alias{RspPageDirective.getTitle}
\alias{getTitle.RspPageDirective}
\alias{getTitle,RspPageDirective-method}


\title{Gets the title}

\description{
 Gets the title.
}

\usage{
\method{getTitle}{RspPageDirective}(directive, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspPageDirective}}.
}
\keyword{internal}
\keyword{methods}
