%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspFileProduct}
\alias{getType.RspFileProduct}
\alias{RspFileProduct.getType}
\alias{getType.RspFileProduct}
\alias{getType,RspFileProduct-method}


\title{Gets the type of an RSP file product}

\description{
 Gets the type of an RSP file product.
}

\usage{\method{getType}{RspFileProduct}(object, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{RspFileProduct}}.
}
\keyword{internal}
\keyword{methods}
