%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspProduct}
\alias{getType.RspProduct}
\alias{RspProduct.getType}
\alias{getType,RspProduct-method}
\alias{getType.RspFileProduct}

\title{Gets the type of an RSP product}

\description{
 Gets the type of an RSP product.
}

\usage{
\method{getType}{RspProduct}(object, default=NA, as=c("text", "IMT"), ...)
}

\arguments{
  \item{default}{If unknown/not set, the default content type to return.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspProduct}}.
}
\keyword{internal}
\keyword{methods}
