%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspString.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.RspString}
\alias{getType.RspString}
\alias{RspString.getType}
\alias{getType,RspString-method}

\title{Gets the type of an RSP string}

\description{
 Gets the type of an RSP string.
}

\usage{
\method{getType}{RspString}(object, default=NA, as=c("text", "IMT"), ...)
}

\arguments{
  \item{default}{If unknown/not set, the default content type to return.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspString}}.
}
\keyword{internal}
\keyword{methods}
