%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspSourceCodeFactory.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{makeSourceCode.RspSourceCodeFactory}
\alias{makeSourceCode.RspSourceCodeFactory}
\alias{RspSourceCodeFactory.makeSourceCode}
\alias{makeSourceCode,RspSourceCodeFactory-method}

\title{Makes a RspSourceCode object}

\description{
 Makes a RspSourceCode object.
}

\usage{
\method{makeSourceCode}{RspSourceCodeFactory}(this, code, ...)
}

\arguments{
  \item{code}{A \code{\link[base]{character}} \code{\link[base]{vector}} of code strings.}
  \item{...}{Arguments passed to the language-specific
     \code{\link{RspSourceCode}} constructor.}
}

\value{
 Returns a \code{\link{RspSourceCode}} object.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspSourceCodeFactory}}.
}
\keyword{internal}
\keyword{methods}
