%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspDocument.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{mergeTexts.RspDocument}
\alias{mergeTexts.RspDocument}
\alias{RspDocument.mergeTexts}
\alias{mergeTexts,RspDocument-method}

\title{Merge neighboring 'text' elements}

\description{
 Merge neighboring 'text' elements by pasting them together.
}

\usage{
\method{mergeTexts}{RspDocument}(object, trim=FALSE, ..., verbose=FALSE)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link{RspDocument}} with equal or fever number of elements.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{RspDocument}}.
}
\keyword{internal}
\keyword{methods}
