%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  processRsp.HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{processRsp.HttpDaemon}
\alias{processRsp.HttpDaemon}
\alias{HttpDaemon.processRsp}
\alias{processRsp.HttpDaemon}
\alias{processRsp,HttpDaemon-method}


\title{Processes an RSP page}

\description{
 Processes an RSP page.
}

\synopsis{processRsp.HttpDaemon(static, ...)}
\usage{HttpDaemon$processRsp(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns (invisibly) the translated RSP document as a single
 \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
