%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DEPRECATED.rsp.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{rsp}
\alias{rsp.default}
\alias{rsp}

\title{Compiles an RSP document}

\description{
 Compiles an RSP document.
}

\usage{
\method{rsp}{default}(filename=NULL, path=NULL, text=NULL, response=NULL, ..., envir=parent.frame(),
  outPath=".", postprocess=TRUE, verbose=FALSE)
}

\arguments{
  \item{filename, path}{The filename and (optional) path of the
     RSP document to be compiled.}
  \item{text}{A \code{\link[base]{character}} \code{\link[base]{vector}} of RSP code to be processed,
     iff argument \code{filename} is not given.}
  \item{response}{Specifies where the final output should be sent.
     If argument \code{text} is given, then \code{\link[base]{stdout}}() is used.
     Otherwise, the output defaults to that of the type-specific compiler.}
  \item{...}{Additional arguments passed to the type-specific compiler.}
  \item{envir}{The \code{\link[base]{environment}} in which the RSP document is evaluated.}
  \item{outPath}{The output and working directory.}
  \item{postprocess}{If \code{\link[base:logical]{TRUE}}, and a postprocessing method exists for
     the generated document type, it is postprocessed as well.}
  \item{verbose}{See \code{\link[R.utils]{Verbose}}.}
}

\value{
  If argument \code{response} specifies a file output, then the
  absolute pathname of the generated file is returned.
  If argument \code{text} is specified, then the generated string
  is returned (invisibly).
}

\examples{\dontrun{
 # EXAMPLE #1: Output random number
rsp(text="A random number in [1,100]: <\%=sample(1:100, size=1)\%>\n")

# EXAMPLE #2: Assign and output random number
rsp(text="A random number in [1,100]: <\%= x <- sample(1:100, size=1); \%>\n")
cat(sprintf("The random number generated by the RSP document was: \%d\n", x))

}}

\author{Henrik Bengtsson}

\section{Postprocessing}{
  For some document types, the \code{rsp()} method automatically
  postprocesses the generated document as well.
}




\keyword{file}
\keyword{IO}
\keyword{internal}
