%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspPage.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{RspPage}
\docType{class}
\alias{RspPage}


 \title{The RspPage class}

 \description{
  Package:  R.rsp \cr
\bold{Class RspPage}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{RspPage}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{RspPage}\cr
extends \link[R.oo]{Object}\cr


 }

 \usage{RspPage(pathname=NULL, ...)}

 \arguments{
   \item{pathname}{A \code{\link[base]{character}} string.}
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:getAbsolutePath.RspPage]{getAbsolutePath}} \tab Gets the absolute pathname to the current RSP file.\cr
 \tab \code{\link[R.rsp:getName.RspPage]{getName}} \tab Gets the (base)name of the current RSP file.\cr
 \tab \code{\link[R.rsp:getPath.RspPage]{getPath}} \tab Gets the path of the directory of the current RSP file.\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save

 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
