%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{RspProduct}
\docType{class}
\alias{RspProduct}


 \title{The RspProduct class}

 \description{
  Package:  R.rsp \cr
\bold{Class RspProduct}\cr

\code{logical}\cr
\code{~~|}\cr
\code{~~+--}\code{RspProduct}\cr

\bold{Directly known subclasses:}\cr
\link[R.rsp]{RspFileProduct}\cr

public class \bold{RspProduct}\cr
extends logical\cr



  An RspProduct object represents an RSP product generated by processing
  an RSP document.
 }

 \usage{RspProduct(object=NA, type=attr(object, "type"), ...)}

 \arguments{
   \item{object}{The RSP product.}
   \item{type}{The content type of the RSP product.}
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:findProcessor.RspProduct]{findProcessor}} \tab Tries to locate a processor for an RSP product.\cr
 \tab \code{\link[R.rsp:getType.RspProduct]{getType}} \tab Gets the type of an RSP product.\cr
 \tab \code{\link[R.rsp:hasProcessor.RspProduct]{hasProcessor}} \tab Checks whether a processor exist or not for an RSP product.\cr
 \tab \code{\link[R.rsp:print.RspProduct]{print}} \tab Prints a summary of an RSP product.\cr
 \tab \code{\link[R.rsp:process.RspProduct]{process}} \tab Processes an RSP product.\cr
}


 \bold{Methods inherited from logical}:\cr
as.data.frame, as.raster, whichVector

 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{classes}
\keyword{internal}
