%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{RspProduct}
\docType{class}
\alias{RspProduct}


 \title{The RspProduct class}

 \description{
  Package:  R.rsp \cr
\bold{Class RspProduct}\cr

\code{logical}\cr
\code{~~|}\cr
\code{~~+--}\code{RspProduct}\cr

\bold{Directly known subclasses:}\cr
\link[R.rsp]{RspFileProduct}, \link[R.rsp]{RspRSourceCode}, \emph{\link[R.rsp]{RspSourceCode}}, \link[R.rsp]{RspStringProduct}\cr

public class \bold{RspProduct}\cr
extends logical\cr



  An RspProduct object represents an RSP product generated by processing
  an RSP document.
 }

 \usage{
RspProduct(object=NA, type=attr(object, "type"), annotations=list(), ...)
}

 \arguments{
   \item{object}{The RSP product.}
   \item{type}{The content type of the RSP product.}
   \item{annotations}{A named \code{\link[base]{list}} of other content annotations.}
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.rsp:findProcessor.RspProduct]{findProcessor}} \tab Locates a processor for an RSP product.\cr
 \tab \code{\link[R.rsp:getAnnotations.RspProduct]{getAnnotations}} \tab Gets the annotations an RSP product.\cr
 \tab \code{\link[R.rsp:getType.RspProduct]{getType}} \tab Gets the type of an RSP product.\cr
 \tab \code{\link[R.rsp:hasProcessor.RspProduct]{hasProcessor}} \tab Checks whether a processor exist or not for an RSP product.\cr
 \tab \code{\link[R.rsp:print.RspProduct]{print}} \tab Prints a summary of an RSP product.\cr
 \tab \code{\link[R.rsp:process.RspProduct]{process}} \tab Processes an RSP file product.\cr
}


 \bold{Methods inherited from logical}:\cr
as.data.frame, as.raster, whichVector

 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{classes}
\keyword{internal}
