%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  RspProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{RspProduct}
\docType{class}
\alias{RspProduct}


\title{The RspProduct class}

\description{
 Package:  R.rsp \cr
\bold{Class RspProduct}\cr

\code{logical}\cr
\code{~~|}\cr
\code{~~+--}\code{\link[R.rsp]{RspObject}}\cr
\code{~~~~~~~|}\cr
\code{~~~~~~~+--}\emph{\code{RspProduct}}\cr

\bold{Directly known subclasses:}\cr
\link[R.rsp]{RspFileProduct}, \emph{\link[R.rsp]{RspRSourceCode}}, \emph{\link[R.rsp]{RspShSourceCode}}, \emph{\link[R.rsp]{RspSourceCode}}, \emph{\link[R.rsp]{RspStringProduct}}\cr

public abstract class \bold{RspProduct}\cr
extends \link[R.rsp]{RspObject}\cr



 An RspProduct object represents an RSP product generated by processing
 an RSP document.
}

\usage{
RspProduct(object=NA, ...)
}

\arguments{
  \item{object}{The RSP product.}
  \item{...}{Arguments passed to \code{\link{RspObject}}.}
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{!} \tab  -\cr
 \tab \code{\link[R.rsp:findProcessor.RspProduct]{findProcessor}} \tab Locates a processor for an RSP product.\cr
 \tab \code{\link[R.rsp:getType.RspProduct]{getType}} \tab Gets the type of an RSP product.\cr
 \tab \code{\link[R.rsp:hasProcessor.RspProduct]{hasProcessor}} \tab Checks whether a processor exist or not for an RSP product.\cr
 \tab \code{print} \tab  -\cr
 \tab \code{\link[R.rsp:process.RspProduct]{process}} \tab Processes an RSP file product.\cr
 \tab \code{\link[R.rsp:view.RspProduct]{view}} \tab Views the RSP product.\cr
}


 \bold{Methods inherited from RspObject}:\cr
print

 \bold{Methods inherited from logical}:\cr
Ops,nonStructure,vector-method, Ops,structure,vector-method, Ops,vector,nonStructure-method, Ops,vector,structure-method, as.data.frame, as.raster, coerce,ANY,logical-method, whichVector


}

\author{Henrik Bengtsson}


\keyword{classes}
\keyword{internal}
