%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{sourceTcl.HttpDaemon}
\alias{sourceTcl.HttpDaemon}
\alias{HttpDaemon.sourceTcl}
\alias{sourceTcl.HttpDaemon}
\alias{sourceTcl,HttpDaemon-method}


\title{Loads the Tcl source for the HTTP daemon into R}

\description{
 Loads the Tcl source for the HTTP daemon into R.
}

\synopsis{sourceTcl.HttpDaemon(static, ...)}
\usage{HttpDaemon$sourceTcl(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
