%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  HttpDaemon.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{stop.HttpDaemon}
\alias{stop.HttpDaemon}
\alias{HttpDaemon.stop}
\alias{stop.HttpDaemon}
\alias{stop,HttpDaemon-method}


\title{Stops the HTTP daemon}

\description{
 Stops the HTTP daemon.
}

\synopsis{stop.HttpDaemon(static, ...)}
\usage{HttpDaemon$stop(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.rsp:isStarted.HttpDaemon]{*isStarted}()}.
  \code{\link[R.rsp:start.HttpDaemon]{*start}()}.
  \code{\link[R.rsp:restart.HttpDaemon]{*restart}()}.
  For more information see \code{\link{HttpDaemon}}.
}


\keyword{internal}
\keyword{methods}
\keyword{IO}
